commit 12

+++ src/main/java/com/proyectosgrt/demo/repository/Solicitudes_Repository.java
@@ -1,11 +1,9 @@
 package com.proyectosgrt.demo.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.proyectosgrt.demo.models.Solicitudes;
 
-public interface Solicitudes_Repository extends JpaRepository<Solicitudes, BigInteger> {
+public interface Solicitudes_Repository extends JpaRepository<Solicitudes, Long> {
 
 }
